+++ ros_ethercat_model/include/ros_ethercat_model/temperature_sensor_state.hpp
@@ -81,7 +81,7 @@ public:
   string getName() const { return name_; }
 
   uint16_t temperature_raw_;
-  uint16_t temperature_;
+  double temperature_;
 
 private:
   string name_;
